完成： category查看，更新 暂搁置
代办：
老师要求的其他需求

+++ src/main/java/News/dao/CategoryDao.java
@@ -11,7 +11,7 @@ import java.util.List;
  * Created by Lee on 2017/5/27 0027.
  */
 @Repository
-public interface CategoryDao extends JpaRepository<Category,Long>{
+public interface CategoryDao extends JpaRepository<Category,Integer>{
 
     //@Query("select c from Category c where c.flag=false")
     List<Category> findByFlagFalse();
